Arduino update and csv files

+++ JetCU_Controller/JetCU_Controller.ino
@@ -290,13 +290,14 @@ int requestTemperature(){
 long requestRPM(){
   //Prompt RPM
   I2c.read(ENGINE_ADDRESS, 0x08, 3);
-  int byte1 = I2c.receive();
-  int byte2 = I2c.receive();
-  int byte3 = I2c.receive();
+  long byte1 = I2c.receive();
+  long byte2 = I2c.receive();
+  long byte3 = I2c.receive();
   long rpm = 0;
   if (byte1+byte2+byte3==248){
     rpm=5.025*byte1+1300*byte2-33.439;
   }
+  //Serial.println(rpm);
   return rpm;
 }
